Move out some Functions from ToSkiaLayer to VL,ImGui.Renderhelper that used by ToSkiaLayer and ImGuiRenderer so one CodeBase for INotifiction Handling (KeyMapping), FontAtlasBuilder, and Scaling
