+++ VL.ImGui.Stride/shaders/ImGuiShader_DrawFX.sdsl
@@ -20,7 +20,7 @@
     }
 
     override stage void PSMain() {
-        streams.ColorTarget = SRgbToLinearPrecise(streams.col) * tex.Sample(TexSampler, streams.uv);
+        streams.ColorTarget = streams.col * tex.Sample(TexSampler, streams.uv);
     }
 
 };
