+++ VL.ImGui.Stride/src/ImGuiRenderer.cs
@@ -88,35 +88,12 @@ namespace VL.ImGui
             if (uiScaling != _uiScaling)
             {
                 _uiScaling = uiScaling;
-                UpdateUIScaling(uiScaling);
+                RenderHelper.UpdateUIScaling(uiScaling);
             }
         }
-
-        static unsafe void UpdateUIScaling(float scaling = 1f)
-        {
-            var style = ImGui.GetStyle();
-            style.ScaleAllSizes(scaling);
-
-            // From https://github.com/ocornut/imgui/discussions/3925
-            //ImGuiStyle* p = style;
-            //ImGuiStyle styleold = *p; // Backup colors
-            //var style = new ImGuiStyle(); // IMPORTANT: ScaleAllSizes will change the original size, so we should reset all style config
-            //style.WindowBorderSize = 1.0f;
-            //style.ChildBorderSize = 1.0f;
-            //style.PopupBorderSize = 1.0f;
-            //style.FrameBorderSize = 1.0f;
-            //style.TabBorderSize = 1.0f;
-            //style.WindowRounding = 0.0f;
-            //style.ChildRounding = 0.0f;
-            //style.PopupRounding = 0.0f;
-            //style.FrameRounding = 0.0f;
-            //style.ScrollbarRounding = 0.0f;
-            //style.GrabRounding = 0.0f;
-            //style.TabRounding = 0.0f;
-            //style.ScaleAllSizes(scale);
-        }
-
         #endregion scaling
+
+
 
 
         // need to be called from VL
