+++ VL.ImGui.Stride/src/Properties/launchSettings.json
@@ -5,7 +5,7 @@
     },
     "vvvv": {
       "commandName": "Executable",
-      "executablePath": "$(SolutionDir)\\..\\..\\vvvv\\\\vvvv_gamma_6.0-0112-g792041305d\\vvvv.exe",
+      "executablePath": "$(SolutionDir)\\..\\..\\vvvv\\\\vvvv_gamma_6.0-0134-gc38de2aca3\\vvvv.exe",
       "commandLineArgs": "--package-repositories \"$(SolutionDir);$(SolutionDir)\\..\\\\Repros\" --nuget-path $(SolutionDir)\\..\\..\\nugets --editable-packages VL.ImGui.Stride -o \"$(SolutionDir)\\VL.ImGui.Stride\\help\\Basics\\HowTo ImGui in Stride.vl\"",
       "workingDirectory": "$(SolutionDir)\\..\\..\\"
     }
